fix: error with count in createProduct lambda
