+++ product-service/bin/product-service.ts
@@ -1,8 +1,9 @@
 #!/usr/bin/env node
 import 'source-map-support/register';
 import * as cdk from 'aws-cdk-lib';
+//import * as iam from 'aws-cdk-lib/aws-iam';
 import * as apiGateway from '@aws-cdk/aws-apigatewayv2-alpha';
-import * as dynamodb from "aws-cdk-lib/aws-dynamodb";
+//import * as dynamodb from "aws-cdk-lib/aws-dynamodb";
 import { HttpLambdaIntegration } from '@aws-cdk/aws-apigatewayv2-integrations-alpha';
 import * as lambda from 'aws-cdk-lib/aws-lambda';
 import { NodejsFunction, NodejsFunctionProps } from 'aws-cdk-lib/aws-lambda-nodejs';
@@ -23,20 +24,20 @@ const createLambda = (name: string, props: NodejsFunctionProps ) => new NodejsFu
     PRODUCTS_TABLE_NAME,
     STOCKS_TABLE_NAME,    
   },
-  ...props,
+  ...props,  
 });
 
-const productsTable = dynamodb.Table.fromTableName(
-  stack,
-  "ProductsTable",
-  process.env.PRODUCTS_TABLE_NAME as string
-);
+// const productsTable = dynamodb.Table.fromTableName(
+//   stack,
+//   "ProductsTable",
+//   process.env.PRODUCTS_TABLE_NAME as string
+// );
 
-const stocksTable = dynamodb.Table.fromTableName(
-  stack,
-  "StocksTable",
-  process.env.STOCKS_TABLE_NAME as string
-);
+// const stocksTable = dynamodb.Table.fromTableName(
+//   stack,
+//   "StocksTable",
+//   process.env.STOCKS_TABLE_NAME as string
+// );
 
 const getProductsList = createLambda('GetProductsListLambda', {
   entry: 'src/lambdas/getProductsList.ts',
@@ -57,21 +58,22 @@ const api = new apiGateway.HttpApi(stack, 'ProductsApiGateway', {
   corsPreflight: {
     allowHeaders: ['*'],
     allowMethods: [apiGateway.CorsHttpMethod.ANY],
+    //allowMethods: apiGateway.Cors.ALL_METHODS,
     allowOrigins: ['*'],
   },
 });
 
-const grants = [
-  productsTable.grantReadData(getProductsList),
-  stocksTable.grantReadData(getProductsList),
+// const grants = [
+//   productsTable.grantReadData(getProductsList),
+//   stocksTable.grantReadData(getProductsList),
 
-  productsTable.grantReadData(getProductsById),
-  stocksTable.grantReadData(getProductsById),
+//   productsTable.grantReadData(getProductsById),
+//   stocksTable.grantReadData(getProductsById),
 
-  productsTable.grantWriteData(createProduct),
-  stocksTable.grantWriteData(createProduct),
-];
-grants.forEach((grant) => grant.assertSuccess());
+//   productsTable.grantWriteData(createProduct),
+//   stocksTable.grantWriteData(createProduct),
+// ];
+// grants.forEach((grant) => grant.assertSuccess());
 
 api.addRoutes({
   path: `/${BASE_URL}`,
