+++ product-service/src/lambdas/mocks.ts
@@ -1,4 +1,4 @@
-const products = new Array(100).fill(null).map((_, index) => ({
+const products = new Array(10).fill(null).map((_, index) => ({
   id: String(index),
   title: `Some Stuff${index}`,
   description: `Awesome description for Some Stuff${index}`,
